@@ -22,8 +22,9 @@ class WindowsVideoPlayer extends VideoPlayerPlatform {
   /// Clears one video.
   @override
   Future<void> dispose(int textureId) async {
-    var controller =
-        VideoPlayerWinPlatform.instance.getPlayerByTextureId(textureId);
+    var controller = VideoPlayerWinPlatform.instance.getPlayerByTextureId(
+      textureId,
+    );
     await controller?.dispose();
     mControllerMap.remove(textureId);
   }
@@ -46,17 +47,25 @@ class WindowsVideoPlayer extends VideoPlayerPlatform {
       // Without the file:// scheme, the IMFSourceResolver API treats the "%"
       // character as a normal string instead of url decoding the path.
       var uri = Uri.parse(dataSource.uri!);
-      controller = WinVideoPlayerController.file(File(uri.toFilePath()),
-          isBridgeMode: true);
+      controller = WinVideoPlayerController.file(
+        File(uri.toFilePath()),
+        isBridgeMode: true,
+      );
     } else if (dataSource.sourceType == DataSourceType.network) {
-      controller =
-          WinVideoPlayerController.network(dataSource.uri!, isBridgeMode: true);
+      controller = WinVideoPlayerController.network(
+        dataSource.uri!,
+        isBridgeMode: true,
+        httpHeaders: dataSource.httpHeaders,
+      );
     } else if (dataSource.sourceType == DataSourceType.asset) {
-      controller =
-          WinVideoPlayerController.asset(dataSource.asset!, isBridgeMode: true);
+      controller = WinVideoPlayerController.asset(
+        dataSource.asset!,
+        isBridgeMode: true,
+      );
     } else {
       throw UnimplementedError(
-          'create() has not been implemented for dataSource type [assets] and [contentUri] in Windows OS');
+        'create() has not been implemented for dataSource type [assets] and [contentUri] in Windows OS',
+      );
     }
 
     await controller.initialize();
@@ -70,15 +79,21 @@ class WindowsVideoPlayer extends VideoPlayerPlatform {
   /// Returns a Stream of [VideoEventType]s.
   @override
   Stream<VideoEvent> videoEventsFor(int textureId) {
-    var player =
-        VideoPlayerWinPlatform.instance.getPlayerByTextureId(textureId);
+    var player = VideoPlayerWinPlatform.instance.getPlayerByTextureId(
+      textureId,
+    );
     if (player != null) {
       return player.videoEventStream;
     } else {
       // send an intialized-failed event
       var streamController = StreamController<VideoEvent>();
-      streamController.add(VideoEvent(
-          eventType: VideoEventType.initialized, duration: null, size: null));
+      streamController.add(
+        VideoEvent(
+          eventType: VideoEventType.initialized,
+          duration: null,
+          size: null,
+        ),
+      );
       return streamController.stream;
     }
   }
@@ -86,56 +101,63 @@ class WindowsVideoPlayer extends VideoPlayerPlatform {
   /// Sets the looping attribute of the video.
   @override
   Future<void> setLooping(int textureId, bool looping) async {
-    var controller =
-        VideoPlayerWinPlatform.instance.getPlayerByTextureId(textureId);
+    var controller = VideoPlayerWinPlatform.instance.getPlayerByTextureId(
+      textureId,
+    );
     await controller?.setLooping(looping);
   }
 
   /// Starts the video playback.
   @override
   Future<void> play(int textureId) async {
-    var controller =
-        VideoPlayerWinPlatform.instance.getPlayerByTextureId(textureId);
+    var controller = VideoPlayerWinPlatform.instance.getPlayerByTextureId(
+      textureId,
+    );
     await controller?.play();
   }
 
   /// Stops the video playback.
   @override
   Future<void> pause(int textureId) async {
-    var controller =
-        VideoPlayerWinPlatform.instance.getPlayerByTextureId(textureId);
+    var controller = VideoPlayerWinPlatform.instance.getPlayerByTextureId(
+      textureId,
+    );
     await controller?.pause();
   }
 
   /// Sets the volume to a range between 0.0 and 1.0.
   @override
   Future<void> setVolume(int textureId, double volume) async {
-    var controller =
-        VideoPlayerWinPlatform.instance.getPlayerByTextureId(textureId);
+    var controller = VideoPlayerWinPlatform.instance.getPlayerByTextureId(
+      textureId,
+    );
     await controller?.setVolume(volume);
   }
 
   /// Sets the video position to a [Duration] from the start.
   @override
   Future<void> seekTo(int textureId, Duration position) async {
-    var controller =
-        VideoPlayerWinPlatform.instance.getPlayerByTextureId(textureId);
+    var controller = VideoPlayerWinPlatform.instance.getPlayerByTextureId(
+      textureId,
+    );
     await controller?.seekTo(position);
   }
 
   /// Sets the playback speed to a [speed] value indicating the playback rate.
   @override
   Future<void> setPlaybackSpeed(int textureId, double speed) async {
-    var controller =
-        VideoPlayerWinPlatform.instance.getPlayerByTextureId(textureId);
+    var controller = VideoPlayerWinPlatform.instance.getPlayerByTextureId(
+      textureId,
+    );
     await controller?.setPlaybackSpeed(speed);
   }
 
   /// Gets the video position as [Duration] from the start.
   @override
   Future<Duration> getPosition(int textureId) async {
-    var controller =
-        VideoPlayerWinPlatform.instance.getPlayerByTextureId(textureId);
+    var controller = VideoPlayerWinPlatform.instance.getPlayerByTextureId(
+      textureId,
+    );
     return await controller?.position ?? const Duration();
   }
 
